ci(release): add conditions for running in ci

+++ release.sh
@@ -2,11 +2,20 @@
 
 set -e
 
-rm -fr dist
-git clone "$(node -pe "require('./package.json').repository.url")" dist
-pushd dist
+IN_CI="false"
 
-git checkout main
+if [ "$1" = "--in-ci" ]; then
+    IN_CI="true"
+fi
+
+if [ "$IN_CI" = "false" ]; then
+    rm -fr dist
+    git clone "$(node -pe "require('./package.json').repository.url")" dist
+    pushd dist
+
+  git checkout main
+fi
+
 git checkout -b dist || (git branch -D dist && git checkout -b dist)
 
 VERSION="$(node -pe "require('./package.json').version")"
@@ -27,10 +36,19 @@ echo "Extracting build to root"
 mv ./build/* .
 echo "Done!"
 
+if [ "$IN_CI" = "true" ]; then
+    git config --local user.email "ci@no-reply.com"
+    git config --local user.name "Automated Release"
+fi
+
 git add .
 git commit -m "build(release): version $VERSION"
-git push --force --set-upstream origin dist
 
-rm -fr ./*
-popd
+if [ "$IN_CI" = "false" ]; then
+  git push --force --set-upstream origin dist
+
+  rm -fr ./*
+  popd
+fi
+
 echo "Success!"
